+++ Private/Convert-ModelToEncoding.ps1
@@ -7,6 +7,10 @@ function Convert-ModelToEncoding {
     )
 
     switch -Wildcard ($Model) {
+        # reasoning
+        'o1' { 'o200k_base'; continue }
+        'o1-*' { 'o200k_base'; continue }
+        'o3-*' { 'o200k_base'; continue }
         # chat
         'gpt-3.5-turbo-*' { 'cl100k_base'; continue } # e.g, gpt-3.5-turbo-16k, -0401, etc.
         'gpt-4-*' { 'cl100k_base'; continue }  # e.g., gpt-4-turbo, gpt-4-32k, etc.