+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ### Unreleased
+- Add model encoding mappings for `o1` and `o3` models.
 - Fix an issue where requests to `*.azure.com` could not be executed when the `ApiType` is `OpenAI`. [#38](https://github.com/mkht/PSOpenAI/issues/38)  
   This allows the use of GitHub Models and Azure ML.
 ```PowerShell